backend/requirements.txt
streamlit
ultralytics
torch
opencv-python-headless
numpy<2.0
Pillow
